requirements.txt
Django==1.8.5
pytz==2015.4
arrow==0.6.0
django-localflavor==1.1

MySQL-python==1.2.5

mailchimp==2.0.9
twilio==4.5.0
icalendar==3.9.0

pylibmc==1.5.0

#stripe==1.19.1
git+https://github.com/jemerick/stripe-python.git@ACH
braintree==3.11.0

gunicorn==19.3.0
raven==5.5.0

# do not upgrade cumulus to 1.0.19, it breaks static files
django-cumulus==1.0.17
python-dateutil==2.4.2

django-sslify==0.2.7

django-debug-toolbar==1.3.2

celery==3.1.18
redis==2.10.3
flower==0.8.3

num2words==0.5.3

Pillow==2.9.0
django-imagekit==3.2.6

unicodecsv==0.13.0

django-auditlog==0.3.1

premailer==2.9.6
django-appconf==1.0.1

django-formtools==1.0

# Django import / export - https://django-import-export.readthedocs.org/en/latest/
django-import-export==0.4.2
